Kanhaiya Lal Bohra/Django Project Setup/django_set/requirements.txt
Django>=5.2.3
djangorestframework
djangorestframework-simplejwt
python-decouple
celery
redis
django-celery-results
python-telegram-bot
